+++ src/crypto/certs.h
@@ -28,35 +28,35 @@ namespace crypto
   {
     return key_pair->self_sign(
       subject_name,
-      subject_alt_names,
-      true /* CA */,
       valid_from,
-      compute_cert_valid_to_string(valid_from, validity_period_days));
+      compute_cert_valid_to_string(valid_from, validity_period_days),
+      subject_alt_names,
+      true /* CA */);
   }
 
   static Pem create_endorsed_cert(
     const Pem& csr,
     const std::string& valid_from,
     const std::string& valid_to,
-    const Pem& issuer_key_pair,
+    const Pem& issuer_private_key,
     const Pem& issuer_cert)
   {
-    return make_key_pair(issuer_key_pair)
-      ->sign_csr(issuer_cert, csr, false /* Not CA */, valid_from, valid_to);
+    return make_key_pair(issuer_private_key)
+      ->sign_csr(issuer_cert, csr, valid_from, valid_to, false /* Not CA */);
   }
 
   static Pem create_endorsed_cert(
     const Pem& csr,
     const std::string& valid_from,
     size_t validity_period_days,
-    const Pem& issuer_key_pair,
+    const Pem& issuer_private_key,
     const Pem& issuer_cert)
   {
     return create_endorsed_cert(
       csr,
       valid_from,
       compute_cert_valid_to_string(valid_from, validity_period_days),
-      issuer_key_pair,
+      issuer_private_key,
       issuer_cert);
   }
 
@@ -66,14 +66,14 @@ namespace crypto
     const std::vector<SubjectAltName>& subject_alt_names,
     const std::string& valid_from,
     size_t validity_period_days,
-    const Pem& issuer_key_pair,
+    const Pem& issuer_private_key,
     const Pem& issuer_cert)
   {
     return create_endorsed_cert(
       subject_key_pair->create_csr(subject_name, subject_alt_names),
       valid_from,
       validity_period_days,
-      issuer_key_pair,
+      issuer_private_key,
       issuer_cert);
   }
 }