+++ tests/infra/consortium.py
@@ -652,6 +652,18 @@ class Consortium:
         r = self.vote_using_majority(remote_node, proposal, careful_vote)
         return r
 
+    def set_service_certificate_validity(
+        self, remote_node, valid_from, validity_period_days
+    ):
+        proposal_body, careful_vote = self.make_proposal(
+            "set_service_certificate_validity",
+            valid_from=valid_from,
+            validity_period_days=validity_period_days,
+        )
+        proposal = self.get_any_active_member().propose(remote_node, proposal_body)
+        r = self.vote_using_majority(remote_node, proposal, careful_vote)
+        return r
+
     def check_for_service(self, remote_node, status):
         """
         Check the certificate associated with current CCF service signing key has been recorded in
@@ -662,11 +674,13 @@ class Consortium:
             current_status = r.body.json()["service_status"]
             current_cert = r.body.json()["service_certificate"]
 
-            expected_cert = slurp_file(os.path.join(self.common_dir, "networkcert.pem"))
+            expected_cert = slurp_file(
+                os.path.join(self.common_dir, "service_cert.pem")
+            )
 
             assert (
                 current_cert == expected_cert[:-1]
-            ), "Current service certificate did not match with networkcert.pem"
+            ), "Current service certificate did not match with service_cert.pem"
             assert (
                 current_status == status.value
             ), f"Service status {current_status} (expected {status.value})"